requirements.txt
requests
python-telegram-bot
pyyaml
wikipedia
pytest
rake-nltk
spacy
responses
mock
coverage
pytest-cov
